refacto: create a class to manipulate vmware architecture based servers
